requirements.txt
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-text-splitters==0.3.4
langchain-experimental==0.3.4
langchain-ollama==0.2.1
langchain-huggingface==0.1.2
sentence-transformers==3.3.1
chromadb==0.5.23
pypdf==5.1.0
PyMuPDF==1.25.2
camelot-py[base]==0.11.0
opencv-python==4.10.0.84
rank-bm25==0.2.2
sentence-transformers>=2.2.0
numpy==2.2.1
pandas==2.2.3
Pillow==11.1.0
requests==2.32.3